requirements.txt
gitpython
langchain
openai
python-dotenv
pyyaml
tqdm
